Remove `Content-Encoding: gzip` header from requests and fix CI
A `Content-Encoding: gzip` header was being added to prover requests, but the
request bodies weren't actually being compressed. Now that we've added
decompression middleware to our backend, this incorrect header would lead it to
parse the body as gzip which would result in a 500 error. This PR removes the
incorrect header, and there will be a follow-up to actually compress the
request bodies.

In order to get CI to pass, I also had to set `CXXFLAGS` to include `-std=c++17
-include cstdint` in order for librocksdb to compile. I ran into the same thing
when running locally. While looking at CI, I noticed that the v2 cache action
is deprecated and about to be removed, so this additionally bumps it to v4.
See:
https://github.blog/changelog/2024-12-05-notice-of-upcoming-releases-and-breaking-changes-for-github-actions/#actions-cache-v1-v2-and-actions-toolkit-cache-package-closing-down

Merges #31

LGTM given by: @katiemckeon