Enable zstd request and response compression
On the response side, this adds the `zstd` feature for the `reqwest` crate
which automatically attaches an `Accept-Encoding: zstd` header to outgoing
requests and decompresses responses with `Content-Encoding: zstd` headers. The
request side was slightly trickier because there isn't built in support in
`reqwest`. I implemented the compression here as a new
`ZstdRequestCompressionMiddleware` which compresses request bodies as a stream
if the body size is at least 512 bytes (a similar minimum size applies in our
backend). I went with this approach because it's compatible with our request
retry strategy and doesn't add extra request time by blocking on the
compression.

A small unrelated change is that we were previously setting a `Content-Type:
application/json` header on all requests, I updated this to only apply when
there is a request body. Our backend tolerates this, but it's more correct to
have it unset when there's no body imho.

Merges #33

LGTM given by: @katiemckeon